move AddressConstants to model
